// Aida_Angular/src/app/shared/models/company-data.ts
export interface CompanyModel
{
    companyId: number ;
    companyName: string ;
    companyUEN: string ;
    incorpDate: Date ;
    address1: string ;
    address2: string ;
    city: string ;
    country: string ;
    pincode: string ;
    mobile: string ;
    email: string ;
    fax: string ;
    gstRegNo: string ;
    industryType: string ;
    status: string ;
    statusDate: Date ;
    phone:string;
}